kbfs.service: fix invalid shell syntax completely failing to work
The main problem is that `keybase config` returns json, which the shell
does not understand, and this was badly papered over by returning a json
key which only contains a value. Unfortunately this still contained
quote characters, which means that fusermount attempted to unmount a
directory *literally* called, e.g. "/home/eschwartz/.local/share/keybase/fs"

The result is that the ExecStartPre command always, without fail,
returns a syntax error, and the kbfs mountpoint will occasionally still
be mounted:

fusermount: bad mount point "/home/eschwartz/.local/share/keybase/fs": No such file or directory

The expected results were either silent success, or in the event of a
non-mounted keybase filesystem, the error message:

fusermount: entry for /home/eschwartz/.local/share/keybase/fs not found in /etc/mtab

Or in the event of a mysteriously missing parent directory
/home/eschwartz/.local/share/keybase, the error message (note the lack
of quote chars in the error message):

fusermount: bad mount point /home/eschwartz/.local/share/keybase/fs: No such file or directory

- The simple fix is to use sed (hopefully available on all systems) to
  remove the quote chars, and re-add them in the shell itself.
- The highly inappropriate not-fix which should never ever be used,
  ever, on pain of losing your right to touch a text editor, would be to
  add an eval on top of everything else this command is already doing.
- The ideal fix would be for `keybase config` to only return json when it
  was asked for json, and the rest of the time return simple key/value
  responses.

Also while we are at it, absolutely nothing in this bash snippet
actually requires bash, so it makes sense to use plain old /bin/sh and
as a bonus avoid a mild startup cost of bash compared to minimal shells
like dash, or busybox ash.

Also while we are at it, it makes no sense to launch fusermount via
`env`. Exotic $PATH lookup rather than relying on hardcoded /usr/bin
only really makes sense if there is a high expectation that the binary
might move its location, e.g. the bash command on BSD systems is in
/usr/local. Whereas fusermount will always be in /bin except on systems
which implement the /usr merge in which case it simply doesn't matter.

And failing that, the command is *already* running in a shell, which
means we *already* have $PATH lookup. So `env` is completely redundant.

Also use $() rather than `` as they are nicer in every possible situation.
See https://mywiki.wooledge.org/BashFAQ/082