folder_branch_ops: delete fresh dir blocks from dirty bcache
When a directory is created, a new dir block is put into the dirty
bcache.  However, if a SyncAll happens before we actually create any
operations in that directory, we were incorrectly leaving the dirty
block around in the cache forever. Besides the memory leak, this was
keeping us from being able to enable journaling after that point.

Issue: KBFS-2825