simplefs: return the last writer of a directory entry
Get this info by using the os.FileInfo.Sys() generic interface, and
introducing a new type of interface that the Sys() return value can
implement.  That way we don't try to get the last writer when using a
local disk file system.

Note: this won't return the last writer for symlink directory entries.
I assume that's ok, but if not we can figure out a different way of
doing this.

Issue: KBFS-2836