tlf_handle_resolve: use the team suffix when checking canonicals
When someone passes a TLF name like "a,b,c (conflicted copy 2018-03-22 #1)"
to FUSE, we need to figure out if it's canonical, but it's hard,
because the "#1" is needed for implicit teams, but during the initial
parsing we don't know yet whether the folder is team-backed or not
without making expensive service calls.

So this PR does makes the right suffix when it knows for sure about
the team-backed-status of the TLF; otherwise, if the user passed in a
"#1", it assumes the user (likely a KBFS server test) already knows
that it should be an implicit team TLF.

Also it wrap the TLFNameNotCanonical error with a stack, because this
was really hard to debug without that.

Issue: KBFS-2652