Systests and AccelerateTeamRekeyd: change work_time_sec to 10 sec (#11008)
Despite team_rekeyd being accelerated in tests, it would still try to
send one message and then postpone work for 60 seconds. So if for some
reason it chooses wrong tribute or the gregor message does not go
through, it's game over for the test. Changing this variable to 3
seconds make it possible for the work message to be retried few times
before giving up.