fix shutdown deadlock between Logout and CachedFullSelf
- see the extensive comment in the code:

resetFullSelferWithLoginStateLock is used to clear the fullSelfer
when a loginStateMu is held. This is trickier than it ought to be because
of deadlock potential. If we try to reach inside our existing CachedFullSelf
and grab the mutex that protects the me object (to clear it), we chance calling
into LoadUser, which can attempt to grab the login state, which would deadlock.
So just do the stupid thing, which is to throw away the existing CachedFullSelf
and swap in a new one.