fix bug in hash computation for 2015-era badWhitespaceChainLinks (#11172)
- It turned out that after recent sigchain optimizations, we were using the sig payload to compute
chain link IDs. This is almost always correct, except for the ~30 links from badWhitespaceChainLinks.
In those cases, perform the "fix" operation which is just to strip off the trailing newline.
- Also fix something that seemed like it could cause bugs but didn't in reality: that we were
potentially mutating payload in ChainLink#fixPayload. So make a copy before the append
- fix a bug in the tests that hid this bug in master --- which is that we need to clear
the link cache to test loading and storing.