app: record installer error text in a new file
Neither `console.log()` and `logger.info()` seems to work for the
install unless you start it from the command line, and this error text
is needed to help debug a rare issue that can't be reproduced from the
command line (I've tried in a script continuously for the past day).

So, until normal logging works in installer.js, write the errors (when
they occur) to a temp file that we can ask users to send us separately
from their `keybase log send`.  Hopefully an admin build user will
find the error before we even need to make this public.