redirector: handle "keybase" appearing in user names
Isolate the checks for the regular run mode and the "fancy" run mode,
so that a "keybase" string appearing in the username doesn't cause the
"fancy" mode check to fail.  E.g., before this commit,
`/Volumes/Keybase (akalin-keybase)` would use, but fail, the regular
check, and the "fancy" check would never be tried.

Issue: KBFS-2872