Remove PUKless members the way we remove crypto
We used to remove PUKless members by looking for invites with
`type="keybase"` and `name=user.uv().percent_form()`. This would break
when user resets and EldestSeqno chanes, or when someone tries to remove
member that was added on old client that didn't create invites with
"uid%seqno" form if seqno=1.