libkbfs: remove separate HasUnverifiedVerifyingKey call
Now that we receive all incarnations of a user from the Keybase
service, and we keep both revoked and pre-reset keys (along with a
timestamp) in `UserInfo.RevokedVerifyingKeys`, there's no need to do
anything special with finalized TLFs.  The client will just
automatically find the right pre-reset key and check the timestamp for
it, without needing a separate `HasUnverifiedVerifyingKey` call.

Issue: KBFS-2902