cr: fix bug with recreated file and subdirectory
A user hit an issue where they deleted a file and several of its
parent subdirectories in the merged branch, but re-created that file
and the parent subdirectories in the unmerged branch.  They did this
for several files, across multiple subdirs.

This triggered a bug related to the fact that action updates to files
actually affect their parent directories, and so parent directories
can be processed twice during the action loop.  If the parent
directory is being recreated, we could end up making two new blocks
for it, and the second one would overwrite any changes that we already
made to the first one, leading to the error.

Instead, make sure we only make new blocks once, while still
preserving the previous correct behavior that a recreated directory
block should always start out empty, regardless of what was still left
in the unmerged branch.

Issue: KBFS-2915