[lib]dokan: Fix truncating files with bash
When a file is opened with FILE_OPEN_IF with Dokan and the file
exists it should be returned, but additionally ErrObjectNameCollision
should be returned.

Fixes KBFS-2812 and Fixes #1494

This ended up requiring quite a lot of restructuring code since
we need to pass up whether a file was created or an existing
one is returned. Doing this on all the return sites would have
been even more messy.