libkbfs: estimate the seqno and timestamp for a revocation
The service gives us a timestamp for when a revocation happened, but
it doesn't yet give us an exact seqno -- it just gives us the last
seqno that the client saw before it did the revoke.  This isn't good
enough to perform a tight check on an MD update by a revoked device,
so this adds some slack into those checks until we can get the right
info from the service (CORE-7924).