Change PacketTag to uint for compatibility with new codec (#11991)
There's ambiguity in the MessagePack spec with regards to how ints
are encoded. Newer versions of ugorji/codec encode numbers of type
int that are > 127 as int{16,32,64}, whereas old versions (like the one we use)
encode them as uint{16,32,64}.

Change the type of PacketTag to uint to be forward-compatible, and add tests.