Fix a latent bug with reembedBlockChanges (#1613)
We relied on decoding a BlockChanges clobbering any existing Info,
which would not be true if the omitempty tag on Info were actually respected
(which the upstream version of our codec library (buggily!) does).

So add a test that catches this, and make it so that we decode into an empty
BlockChanges and then assign.

Also, remove omitempty from Info in BlockChanges. We can't turn it back on
until old clients with the buggy behavior all disappear.