kbfsedits: add a class that can construct per-TLF edit histories
This introduces `kbfsedits.TlfHistory`, which receives raw
notification strings from a caller for each writer in a TLF, and
constructs the edit history from each writer (up to
`maxEditsPerWriter` edits each), up to `maxWritersPerHistory` writers
per TLF.

The collected sets of per-writer edits in the history are sorted in
reverse order, by the most recent revision from each writer.  Within
each set, a writer's edits are sorted in reverse order by
revision.

The constructed history only includes creates and modifies of files
(no directories or symlinks).  If a file was modified and later
renamed, the notification for the modify shows the new file name.  If
a file was modified but later deleted, it doesn't appear in the
history.  Each writer only has the most recent modification or create
for each file.

For performance reasons, it's possible that not all deletes or renames
will be processed, because they might have been done by a writer prior
to their `maxEditsPerWriter` edits, and so those won't be reflected in
the modifications of any of the other writers.

The idea is that each `libkbfs.FolderBranchOps`, subscribe to the
kbfs-edits chat channels for the TLF's notifications, and fill in
these histories with notifications as needed.  Then a new,
yet-to-be-written object in `kbfsedits` will collate together a
complete history for the logged-in user, using the individual
`TlfHistory` instances corresponding to the most-recently-written
TLFs.

NOTE: This doesn't yet take journal revisions into account.  That
support will come in a future commit so it's easier to review.

Issue: KBFS-2996