fix a bug where we weren't respecting symmetric key context
This was causing different applications to derive the same symmetric
encryption key. Currently only the stellar code is affected, which isn't
yet publicly launched. We believe we can replace the bad boxes in the
employee accounts that have issued them, as part of switching this code
over to a proper context string.