Fix avatar mkdir
It used to be called in a place where GetCacheDir returned wrong answer,
because it was before commandline was parsed. So e.g. if development
build was launched, even if --run-mode=prod was passed, it would not be
seen yet and ".cache/keybase.devel/avatars" dir would be created. Later
on, GetCacheDir would start returning expected cache dir (".cache/keybase/"),
and avatar loader would not be able to find "avatars/" subdir.