cr_chains: don't revert rename op if it was just created
If an entry is created and then renamed within a single journal
update, `crChains` ends up collapsing the [`createOp`,`renameOp`] into
a single `createOp`, but with `createOp.renamed = true`.

This confused the `revertRenames` function, because it couldn't find a
matching `rmOp` for the given `renameOp`.  But that's valid, so we can
just leave that `createOp` un-reverted.  Otherwise, we'll have a
`renameOp` without a valid `oldFinalPath`, and we'll crash.

Issue: keybase/client#12817