Only show @here and @channel in team chats
This commit also handled a very annoying Flow issue. After adding the
prop `isTeam` to `mention-hud-container` Flow was endlessly erroring
that `isTeam` is missing in object literal `A`.

The solution was to add an `index.js.flow` file to the directory to
force flow to understand the module level props.

Cause of the original error is not totally known.