prev_revisions: avoid panic when count is about to overflow
jzila called it.

Some user reported an issue where the count can reach 255.  I can't
repro it yet, but I suspect it has to do with removing revisions from
the front of the list when running conflict resolution, and thereby
letting the entries at the end of the list build up their count.