Remove CircleCI cache (#1726)
(This is a port of a PR from client.)

The main problem was that
~/.gradle/caches/transforms-1/files-1.1/keybaselib.aar is a
cached directory that stores every new keybaselib.aar build.

But we don't really need to be caching anyway. It seems that CircleCI already
does some sort of download caching. So just nuke caching entirely.

Also fix build breakage when there is no cache -- we'd previously run into
it periodically.