dir_data: fix lookup semantics
Before the previous commit that got rid of `deCache` in favor of
`dirData`, `folderBlockOps.getDirtyEntryLocked` used `blockLookup` as
the request type when reading blocks.  This was to make sure we did
not release `blockLock` when making network accesses to fetch blocks,
like we normally do with `blockRead`.  It's a minor optimization, but
if we do that while creating an entry, we could end up in a bad race
situation where two Nodes are created for the same entry name.

So revive those old semantics so that race doesn't come back to haunt
us.

Issue: KBFS-3302