first implementation of the teamchain auditor (#13544)
- the goal of the teamchain auditor is to prevent the server from pulling off an "odd/even" attack. That is, when showing team acme to
alice, the server sends even merkle roots, that show view A of the team. When showing team acme to bob, the server sends odd
merkle roots, that show view B of the team. Where A and B might not agree at all. The server can hide these patterns in the
merkle tree in arbitrary ways, with the goal to get alice and bob to see two different ideas of the same team, without forking
the merkle tree globally.
- our approach is for alice and bob to make random "probes". the probe the team before the team was created, the time after the
team was created, and ensure a consistent view.
- after 1000 new merkle roots, the audit repeats, with alice and bob idependently auditing the recent roots.
- for now, we've hardcoded parameters, but we'll likely want to tune them in the future.
- write all audit results to disk cache.
- merge by hand since github is down
- also @mlsteele might have further comments