folder_block_ops: better unref management for dirty dirs
When undoing a dir entry operation, we need to un-dirty any blocks
that had been dirtied for the first time, by removing them from the
dirty block cache.

Also, when deleting big dir trees, make sure the unrefs propagate all
the way down to the remaining parent, so they can be counted in the
final resolution.

Finally, move the unref accounting to the top of the prep loop, to
make sure the first passed-in block gets properly accounted for.

Issue: KBFS-3303