Change how DeviceWrap works w.r.t device switching, cleanup selfprovisioning (#13406)
Added Rollback method to ConfigWriterTransacter, that will reload
config back from file or restore empty state if it was transacting on
a new file. This provides transaction semantics for jsonFileTransaction 
where caller can cancel operation and restore previous state.

Encapsulate active device switch and config write into separate method
of DeviceWrap engine. This way it's clear for the caller where the
"point of no return" is with regards to provisioning.