Fix misencoded RPCs after a disconnect (#13597)
Our RPC library takes a list of arguments, but our RPCs only ever use a single
element with keyed arguments. Therefore, we have to wrap the argument object
in an array, and we were doing this in TransportShared.invoke.

On disconnect, the arguments to invoke were being stored in a queue,
which is then flushed to call invoke again. Unfortunately, this results in array-wrapping
again, which the service doesn't expect. This is what was causing the strange
RPC decode errors on Windows.

Move the wrapping a few levels up, and add a regression test.

Fix types in a few places, and add more types.