cr: ignore `changeOriginal` call between two same pointers
I've been unable to get good logs that reproduce this issue, and it
seems safe to just ignore the case where CR is trying to change an
original between two identical pointers.  This probably indicates
something like the device putting an MD, but failing before it finds
out that the put succeeds. Then later it tries to resolve the conflict
against itself.  In theory, we have other protections against this,
and it would be nice to figure out why it's causing problems in this
case, but it's probably best to paper over it for now and just give
the user a smoother experience.

It's possible that when the CR finally succeeds, it will end up with a
bunch of conflict files, which might get the user's attention quickly
and entice them to send logs soon (thus capturing the original source
of the problem).

Issue: KBFS-2946