Improve logging on Android (#13747)
Also log startup events on Android. This is to get a better idea of
bottlenecks on the native side.

Make NativeLogger methods not throw, and also take an optional exception. This
makes it a drop-in replacement for Log.i, etc. Port most things to use NativeLogger.

Log any thrown exceptions via NativeLogger.