libgit: add a git browser that doesn't require a checkout
`Browser` uses the gogit `Tree` instance to navigate directories and
read files directly from a dotgit directory, without requiring an
explicit checkout.

This will be used in future commits as the basis for the new
`.kbfs_autogit` directory.

The test provided is pretty simple but covers the basics.

Issue: KBFS-3421