elimite a race condition around folder creation
when we hear from journal after a new folder is created, and beofre we
get the result of the folderList RPC call triggered by editSuccess, we
don't know yet from pathItems that it's a folder. Without this commit we
might show user the upload banner since the type is 'unknown'. This
commit makes the upload container look at state.fs.edits as well to
eliminate this race.