folder_branch_ops: implement Node.GetFS-based, read-only browsing
With this commit, `folderBranchOps` uses the new `Node` methods
`GetFS()` and `GetFile()` (when they return non-nil instances) to
provide data for the read-only KBFSOps methods (`Stat`, `Lookup`,
`GetDirChildren`, and `Read`).

It also adds a new entry type, `FakeDir`, that can be returned by the
existing `Node.ShouldCreateMissedLookup()` method.  This is to avoid
having to actually create the root of the browseable file system
(i.e., `.kbfs_autogit` in future PRs), so it can be used without
requiring write access to the TLF.

Because the node cache requires valid `BlockRef`s for each node, we
need to generate fake `BlockID`s for each fake entry.  Each `BlockID`
only needs to be unique within a TLF, since they never touch the block
cache in memory or on disk.  So, we simply generate a normal block ID
by passing the (unencrypted) path within the TLF to the block ID hash
function.

Issue: KBFS-3424