libgit: support for browsing branches
With the root of an autogit repo, you can now navigate to the branch
of a version by using `.kbfs_autogit_branch_<name>`.  Branch names can
contain slashes, so you can stack these.  For example, to navigate to
a branch a/b/c in repo "test", use this path:

.kbfs_autogit/test/.kbfs_autogit_branch_a/.kbfs_autogit_branch_b/.kbfs_autogit_branch_c/

It's a bit unwieldy, but it will be fix for GUI-based browsing, and I
don't see a good alternative for CLI-based browsing, besides
shortening the prefix somehow.

If the branch names doesn't start with "refs/", then "refs/heads/" is
prepended to it.  But e.g. tags can be shown with a full path like
"refs/tags/v2.6.0".

Issue: KBFS-3426