node_cache: add and implement AllNodeChildren function
This function returns all the children of the current node.  It will
be useful to figure out which autogit nodes need to be invalidated,
after a repo is updated.

This implementation iterates through all nodes, and back through all
parents, to figure out which nodes to return.  A better implementation
might track the a node's children directly, but that will be a bit
tricky to get right, so let'd try this for now and see if it's good
enough.

Issue: KBFS-3427