autogit: cache and reuse `*Browser` objects
Instead of creating a new `*Browser` for every wrapped Node, have
`AutogitManager` create them on-demand and cache them for future use.

Previously, when the browsers were stored in the node objects
themselves, they'd get cleaned upon kernel node garbage collection.
With this new structure, we use an LRU cache of a limited size to
control the memory.

The browsers themselves are stuck at a particular commit, so they need
to be cleared out everytime the underlying repo changes.

The improved the performance of reading an entirely-disk-cached
~30-entry directory from 6.35s to 0.66s on my machine.

The main downsides are that:

* Since the size of the LRU cache is independent of the kernel cache,
  we might end up throwing away browsers for nodes that are still in
  the cache, and they have to be re-made again on the next access.
  But that doesn't seem too likely in practice.

* Previously we were using the caller's context objects when
  constructing the browsers, which meant we would have the caller's
  log tags for debugging purposes.  This way uses the log tags of
  whoever first created the cached browser, for all subsequent calls.
  This is a bummer, but to fix it we'd need to do surgery in go-git to
  be able to update the storage layer of their `Tree` objects on the
  fly, which isn't possible right now.  We can revisit that later if
  it seems useful.

Issue: KBFS-3474