cache TeamName -> TeamID resolutions as a side effect of fetching favorites
- when a user does an `ls` in `/keybase/team`, it maps to a GetFavorites call,
and then, in common cases, an immediate call tio Resolve3 on every team in the
list. With this PR, we plumb the team ID back from the server (it was already being
sent down) and then prime the resolver cache with these resolutions. It should hopefully
immediately return in this case, rather than incuring a server round trip.