simplefs: operation for resetting TLFs
This allows the user to reset their TLF by generating a new TLF ID for
the implicit team, with the epoch incremented.  It clears the old TLF
out of the running instance.  The new TLF ID is written to the
implicit team sigchain, and overrides the old one.

It can only be done AFTER a Keybase admin does a server-side reset, so
we can track which TLFs are being reset.

Issue: KBFS-3545
Issue: keybase/client#14319