folder_branch_ops: commit MDs to disk after reading/prefetching
FBO nows commits staged MDs for synced TLFs to disk once one of the
following conditions is met:
  1) The new revision is fully prefetched.
  2) The new revision has been read by the user.

In addition, FBO won't even show a downloaded revision to the user
until at least the root block for the revision has been prefetched.

All of this is to ensure that a synced TLF contains at least some data
if the user suddenly loses connectivity, or restarts in offline mode.

Issue: KBFS-3505