folder_block_manager: background cleanup of the sync cache
A synced TLF needs to be able to clean up old cached blocks, even if
it is offline when the block is unreferenced by a different device.
So we track the last unref'd revision for each synced TLF, fetch each
merged revision since the last one, and delete all the unreferenced
blocks from it.

Issue: KBFS-3507