quota_usage: launch background fetch if there's a timeout
We put a pretty short 500ms timeout on fetches when data is already
cached on disk, to provide quick quotas when offline.  But if we do
timeout, we might as well launch a background fetcher with no timeout,
in case we're online but our network latencies are poor.

Suggested by songgao.

Issue: KBFS-3506
Issue: #1873