folder_branch_ops: only commit MDs to disk on a read
This fixes a test flake, where MDs could be commited during the part a
`SyncFromServer` call that did a `SyncAll`.  The test's timing depends
on the commit call happening only while fetching the updates from the
server.

And committing only on reads is what we want anyway -- once the
changes have been seen by the user, they should be committed.  But
writes don't count for that (and they'll be producing a new revision
anyway).