linux: respect XDG variables in the systemd services
keybase.service should respect the XDG_CACHE_HOME when creating its
logfile, but the service hardcoded ~/.cache.

This really should use systemd's '%C' specifier, but in order for
systemd to know the right custom cache directory it needs to be
configured before the user session manager has started, i.e. in
.pam_environment, and users are far more likely to use their bashrc for
this. At that point it is too late for systemd to be natively aware of
it. We cannot just directly use XDG_CACHE_HOME either, as we cannot
guarantee the variable is set at all, and the spec mandates we fallback
on the default value rather than expanding empty.

Cater to the majority of users and do yet more fancy variable exports,
pushing these into the EnvironmentFile for our services to guarantee
that we respect them.