disk_block_cache: the caller of Get/Put specifies the cache type
In anticipation of partial syncing, where the `*folderBranchOps`
instance is the only place that knows whether a given block needs to
be in the sync cache or not, this commit removes usage of
`config.IsSyncedTlf` from the disk block cache.

Instead, it plumbs through a "cache type" with every `Get` and `Put`
call.  For `Get` calls, if the block isn't already in the desired
cache type, it will be moved there (which will be useful when a path
is marked as syncing after some of its blocks are already in the
working set cache).

This commit plumbs the cache type down from the places where the
get/put calls originate from.  Future work will need to make decisions
in those places about whether a particular block goes into the sync
cache; for now, it just uses `config.IsSyncedTlf` there.

Issue: KBFS-3523