ops: show a special error on timeouts for unsynced TLFs
If we are about to return a timeout error to the user when they are
reading data from an unsynced TLF, return a special error type that
lets the GUI show a message explaining that the TLF is unsynced.

Unlike in the case of archived data (where we first check if we're
disconnected and then immediately return the error), in this case we
give the system time to recover from the connectivity issue during the
operation, to avoid spurious notifications that might be recoverable.

Issue: KBFS-3585