Add engine.reset on KeybaseEngine.destroy (droid)
There's a subtle bug where if a thread is blocked on `readB64` from the engine when we call `shutdownNow` on the executor, the thread won't finish and we enter that timeout condition. This results in a broken app since the engine replied to a message, but it was dropped because this thread was orphaned and had no way to communicate back to JS.

With `Keybase.reset()` we finish the block `readB64` call so that this thread can propely shut down. And the new thread that comes up will get the message and be able to forward it properly to JS.

Now we never hit that timeout, so we should not see that error msg again.