prefetcher: fix memory leak and possible hang when skipping fetch
Consider this simple scenario:
* Blocks a -> b -> c
* User requests block a, which is fetched with a
  `BlockRequestWithPrefetch` action.
* Block a is fetched, triggering a prefetch of b with a
  `BlockRequestPrefetchTail` action.
* When the b request is handled by the main prefetcher loop, it skips
      further prefetching because the action says it should stop.

That's all correct, but then the prefetch requests for b and a both
stay in `p.prefetches` forever!  If a is never requested again, it is
never removed from that map, causing a memory leak (including a
pointer to the assembled block).

Moreover, if someone has requested the prefetch wait channel for a and
is waiting on it, they'll wait forever since it'll never be closed.
The correct behavior is for the channel to be closed whenever all
scheduled work is complete, even if the status is not supposed to be
`FinishedPrefetch`.

So instead, we now make sure to cancel a request that we're skipping,
as long as some other previous request is not already prefetching it.

Furthermore, if the block is a tail block, we can mark it as finished
right away and complete the prefetch, instead of just skipping it.

This also fixes up some action-propagation issues to make sure that a
new, upgraded action always overrides an existing lesser action.

These changes alter the timing of the partial sync test a bit, and so
the test is now simplified to only check the statuses of the blocks
that should definitely be finished prefetching.

It also simplifies the prefetcher tests by waiting on specific
prefetches to finish, rather than waiting for the whole prefetcher to
shutdown.