folder_branch_ops: mark and sweep for partially-synced TLFs
Partially-synced TLFs need to clean up the sync cache in the
background, in the cases where a) some previously-synced path is now
unsynced manually by the user, or b) the user moves some
previously-synced directory out of a synced path.  In the latter case,
we don't have an explicit config action from the user, so it needs to
be done periodically in the background.

This adds a new background goroutine to FBO for partially-synced
folders, that triggers a mark-and-sweep operation either when a path
is removed from the sync config, or on a periodic timer (once an
hour).  It iterates through each synced path and marks each block that
should still be synced with a tag.  After that, it removes all blocks
from that TLF that aren't marked with that tag from the sync cache.

This process is stopped whenever the latest merged MD is updated.

Issue: KBFS-3644