block_retrieval_queue: fix bug failing to use mem-cached blocks
If a block was in the memory cache, but had no prefetch status in our
preferred on-disk cache, we drop down to do a full get of the block
(which moves the block into our preferred cache).  But in that case we
were leaving `block` filled in (from the memory cache), which caused
an error when trying to decode the encoded-disk version of the block
again, and this caused the block to never be read from the disk cache.