cr_actions: only apply `renameUnmergedAction` to correct ops
If there are more than one unmerged action in a directory, we are
calling all the actions related to items in that directory for each of
the unmerged pointers.  For `renameUnmergedAction`, this had the bad
effect of clearing out all the ref blocks in `syncOp`s and changing
the file pointer in `setAttrOp`s.

Instead, make sure that only happens for the exact unmerged pointer
chain associated with that action.

Issue: KBFS-3770