always refresh path metadata at mount or whenever path changes
Bug before this change: when user gets to a TLF through a link in chat,
we'd know it's a folder and wouldn't load metadata for it. But we
actually don't have a PathItem for it, thus have no info on whether it's
writalbe. So just always load it on component mount, and in update
whenever path changes.