folder_block_manager: fix and simplify quota reclamation
Previously we would give up doing QR if there were more than 100
recent revisions that were less than 2 weeks old.  This is clearly the
wrong thing to do.  I think we did this previously due to historical
limitations of old data formats (i.e., not having the `LastGCRevision`
field, and not being able to tell when an MD revision was made).

These limitations are ancient now, and it's notably preventing QR on
some folders.  So fix it by counting MDs forward from the
`LastGCRevision`, instead of backward (which could result in an
arbitrary amount of work).

Issue: KBFS-3597