root_metadata: initialize TLF with LastGCRevision=1
For TLFs that are older than the `LastGCRevision` field, the QR
process has to walk backwards to find the last GC op.  But for new
folders, we shouldn't have to pay this price when doing the first QR.
Since revision 1 will never have any unrefs, it's safe to mark it as
GC'd from the very beginning, to avoid that walking backwards.

Issue: KBFS-3597