Fix seqno state on error, plus check amounts
There was a bug that when SubmitPayment or SubmitRelayPayment
returned an error, the pending tx wasn't removed from WalletState.
This causes seqno issues after that...

Also, this adds some more checks for low payment amounts up front.
Since we are predicting seqnos so much now, it doesn't make sense
to burn one on a low amount payment that we can detect easily.