md_ops: don't fetch 10 MDs at a time when checking copied MD
Checking 10 at a time was intended to be an optimization, in case
there were multiple MD revisions in a row with a copied data field.
But in practice, this led to a lot of problems because if any of those
10 revisions had its own copied field, the process would recurse and
another 10 would be downloaded, etc.  Even worse, when we introduced
merkle-checking, if any of those 10 were signed by a revoked device,
we would have to go through the full merkle check.  And on a new
device, we might not have the keys for that, and it just gets worse
and worse from there.

So instead, let's just minimize the amount of that kind of
verification by only getting one MD at a time, as needed.  In the most
common scenario, this will only have to fetch one MD, and we'll save
both CPU and bandwidth.

Issue: KBFS-3805