cr: store pointers in the `fileBlockMap`, not full blocks
Change the map that was storing the top block of each changed file in
a directory, into just storing the file pointer for each changed file
in a directory.  The blocks themselves are already stored in the dirty
block cache, so no need to keep them in memory in this other data
structure.

Issue: KBFS-3679