prefetcher: more control for the reschedule test
The reschedule test requires precise timing to get rid of flakes,
because it needs to set the cache limits at exact times during the
prefetch processing.  To get the right level of control, add a second
test channel to the prefetcher, to let us know when an operation is
done being processed.  This exposed other bugs in the way the prefetch
status is updated in the block cache, which are also fixed here.

Issue: KBFS-3779