mount_test: sync fs2 before accessing files via mnt2
Not syncing used to work fine I think, because the first read would
actively sync the folder.  But now with the edit notification stuff,
the folder could be initialized at an earlier revision before an
explicit read via the mount is done.  Thus, we need explicit syncing
to make sure it's up-to-date.

Issue: KBFS-3758