md_util: re-embed block changes when needed, if they're missing
A previous commit to load the block changes into a copy of the MD (to
fix a data race) made other races more likely to happen.  In
particular, if conflict resolution or pointer archival read the
"original" MD from the cache, they would try to operate on it and
would never see the re-embedded block changes, even though they depend
on them for correct operation.

So if we ever hit that race, manually re-embed the block changes
before using the MD.

Issue: KBFS-3836