test: add regression test for setmtime(dir)+write(file)
Before the fix for KBFS-3770, this was broken because the syncOp for
the file would be attempted to be applied for the conflicting
directory (which is a symlink), and it would fail because it expects
the entry for a syncOp to have a BlockPointer (but symlinks do).

When I revert the fix for KBFS-3770, this test fails.

Issue: KBFS-3813
Issue: keybase/client#15699