md_ops: wait for MDServer to be initialized
In the init code, `MDOps` is set before `MDServer`, and so there's a
race condition where something (e.g., the disk cache initialization
code that triggers when the keybase service connection is established)
can try to use the MDServer before it exists.

So just add a simple sleep for the very rare case where this could
happen. This doesn't seem like it's worth the code complexity of
something that would get notified exactly when the `MDServer` is set.

Issue: KBFS-3587