folder_branch_ops: retry md request when forcing a fast forward
When a user logs out, and then logs back in, they need to both:

a) re-authenticate to the mdserver (and bserver)
b) re-initialize all their private/team TLFs (which had their MDs
   cleared on the logout) via fast-forwarding

These happen simultaneously, and it's possible to have a race where
(b) happens before (a) for some TLFs.  When that happens, the
fast-forward hits an unauthorized error from the mdserver.  This
renders the TLF inaccessible until the next logout via the mountpoint,
because `fbo.head` isn't set and only `GetRootNode()` will set it in
that case, but the file system already has the root node cached and
won't try to get it again.

Instead, when receiving the unauthorized error, just try again for a
few seconds and see if it gets better.

Issue: KBFS-3802