libgit: treat `OfflineUnsyncedError` as timeout when repo-cleaning
`folderBranchOps` now turns a `DeadlineExceeded` error into this kind
of error for unsynced TLFs.  When cleaning repos during a git
operation, we're just trying to remove as many old files as we can
within the deadline window, so we want to ignore those kinds of errors
in that case.

Issue: KBFS-3864