kbfs_ops_test: wait for indirect ptr blocks in partial sync test
In this test, the root node and `a` node are indirect blocks, and each
have multiple child blocks.  To get the right counts in the sync
cache, we need to wait for all of them to complete.  We can't just
wait for the blocks to prefetch, because `a` has an entry `b2` that is
never itself prefetched.

Issue: KBFS-3840