Stellar settings: be aware if we can send tx, do not show "Set Inflation" button if not (#15756)
Present the distinction between "funded accounts" and "not funded", as well as
"can submit tx" accounts (that are both funded and have enough for a tx fee) to
the frontend. Use the latter to hide "Set inflation" button if this account
can't set any inflation yet because it's fresh (not on the network yet), or has
not enough funds to submit setInflation tx.