folder_branch_ops: avoid deadlocks when chat monitoring is stopped
If the chat monitor loop gets an error, it shuts down because it
doesn't know how to recover from them.  (For example, an unparseable
TLF handle, or a temporary network error.  Hopefully soon the service
will give us a typed error for the latter so we can backoff and retry
it...)  But if it gets edit messages after that, it could still be
sending them into the fixed-size `editActivity` channel, and
eventually cause a deadlock when that channel fills up.

Instead, introduce a new channel that is created when monitoring
starts, and is closed when monitoring ends, and avoid trying to send
messages if monitoring is not active.

Issue: KBFS-3831