block_retrieval_queue: put block in cache when prefetching is off
If prefetching is off (like on mobile), we were never storing fetched
blocks in the in-memory cache.  They would still be stored in the disk
cache, but in some rare cases that cache can't be initialized on
mobile, which could lead to the same block being downloaded over and
over again.

Issue: KBFS-3891