folder_block_manager: cancel all prefetches before cleaning cache
When `folderBranchOps` applies the changes from a new revision, it
goes through each updated pointer, and cancels prefetches for the old
version of the pointer.  But the pointers aren't ordered in anyway, so
this scenario is possible, when block A points to block B:

1) Cancel prefetch for B.
2) Prefetch for A starts, which schedules a prefetch for B.
3) Cancel prefetch for A.
4) Cache-cleaning completes.
5) Prefetch for B starts, and block B is put into the cache after it
   was cleaned.

To fix this, let's do a second round of prefetch-cancellations before
cleaning cache, and wait for each one individually to complete.  These
two cancellation loops together should be enough to ensure that the
block won't end up back in the cache.

Also, cancel prefetches for unreferenced (not just updated) blocks,
which was an oversight in the existing code.

Issue: KBFS-3784