Zero _ignoreScrollRefCount before handling page up/down (#16071)
We used it whenever we programmatically scrolled, and wanted the next
scroll event to get ignored. However, it turns out that changes to scrollTop
don't correspond 1:1 to scroll events.

In particular, it's possible for the ref count to get "stuck" at > 0, in which case
hitting page up would trigger a scroll event that would get ignored, causing
isLockedToBottom to remain true, which would then cause onResize to snap
back the scroll event and re-increment the ref count.

So zero it out before we handle page up/page down, since we're handling
a user-driven scroll event.