cr: use original ptr when finding path for created chain
It's incorrect to use `mostRecent` when looking in `createOp`s for the
parent directory of a newly-created chain, because the `createOp`
always references the original pointer, and if the node in question
was later modified in a different batch, the original and most-recent
pointers would be different, by definition.

Adds a test that reproduces the "index out of range" error experienced
by the user in KBFS-3915 without this change.

Issue: KBFS-3915