keybase_service_base: pass offline availability to `Resolve3`
This pass `BEST_EFFORT` offline availability mode to `Resolve3`,
whenever the TLF path or ID in question is being synced.  This tells
the service that it is ok to serve a stale answer from the cache, if
the API server can't be reached in a timely fashion over the internet.

Issue: KBFS-3918