Filter arguments for `chat api-listen` (#16364)
Add --filter-channel and --filter-channels arguments to `chat api-listen` for
chat message filtering. Both take a JSON encoded channels specified similarly
to how `keybase chat api` represents them, but --filter-channel takes one
channel JSON object and --filter-channels takes a list.

The command still works as before by default - if no filters are specified, all
messages are relayed. If any filters are used, message is relayed if it matches
one of them.

"topic_type" and "topic_name" are optional, and when they are not specified,
the filter will match any topic_type (so CHAT or DEV in this case). If
"topic_name" is not specified, messages from all channels will be relayed.