folder_branch_ops: don't error on initial MD with no root block
This could indicate that we're initializing with an old cached MD, for
which the blocks have already been GC'd and can't be fetched.  In that
case, we still want to continue onto `setHeadLocked` so that we
register for MD updates to get up to speed, but we don't want to
commit the MD to disk if it happens to be something we fetched.

Issue: KBFS-3946