folder_branch_ops: transform root block errors into offline errors
If the device is trying to access an unsynced TLF via an out-of-date,
cached MD, the root block may have been GC'd.  We want to expose a
slightly nicer error to the user in that case, so the
`OfflineUnsyncedError`, which on desktop causes a popup with more info.

The head should be fast-forwarded once the FBO registers for and
processes updates.

Issue: KBFS-3946