libkbfs: move all data-related code into a new `data` package
The goal of this refactor is to shrink the size of libkbfs, and to
contain most of the logic around block data into its own package,
which should make it easier to find/understand and more documentable.
In addition, it should make it much easier to split other big chunks
of code out of libkbfs, since the new `data` package provides most of
the major dependencies of that other code.

Here's the most important stuff that was moved:

* `BlockPointer`
* `DirEntry`
* `EntryInfo`
* `BlockInfo`
* `BlockCache`
* `DirtyBlockCache`
* `blockTree`
* `fileTree`
* `dirTree`
* `BlockSplitter`
* `path`
* `Ready()`
* `PrevRevisions`

Of course, the moved structs also needed to be exported in some cases,
and to export all of their methods.

In addition, there were some test dependencies reliant on this code
that I didn't want to move into the proper `data` package (because
then they'd get sucked into the production binary), so I made a few
minor, testing-only packages for them as well.

* libkey/test: provides the EmptyKeyMetadata struct used by tests.
* test/clocktest: provides the `TestClock` used by tests.

Finally, some of the block processing code depended on some crypto
functions that were in `CryptoCommon`, but were just wrappers for
other functions.  So I removed those `CryptoCommon` functions in
preference for just calling those functions directly.  For example,
`MakeTemporaryBlockID()`.

I also moved block padding/de-padding into the kbfscrypto library,
instead of being in the `CryptoCommon` struct for no reason.