libkbfs: extract all identify and tlfhandle code into new packages
This creates two new kbfs subpackages, consisting of existing code
moved around:

* idutil: Basic data structures, interfaces, and helper code for
  dealing with identity data for users and teams.
* tlfhandle: The data structure for "Handles" to top-level folders
  (TLFs), which represent an identifier for each TLF, containing all
  the user or team IDs associated with the it.

Not only are these decent cleanups on their own, they will enable
further refactoring of code that uses TlfHandles into other new
packages (which is a lot of them).

Also I fixed up the README to point to all the new packages we've
added since it was written.