libgit: register repo nodes for updates on branches
If the user _only_ looked at a non-master branch, we were not
registering with the `AutogitManager` to get updates for the
repository's root node.  The `AutogitManager` already makes sure to
only register with KBFS once for each TLF, so it's safe to register
with the manager for each different branch.