Prime secret store before provisioning (#16647)
This PR adds a PrimeSecretStores function that check if the secret store is
functioning - that is, it tries to store, retrieve, and then cleanup a random
secret for randomly generated username.

Calls to this new check were inserted in loginProvision, SelfProvisionEngine,
and SignupEngine. The engines are stopped early if they discover that the
secret store does not work. Normal login (LoginProvisionedDevice) is not
protected, so it will still work, just not remember user's passphrase in case
we can't use secret store.

This is important for NOPW sign up, in which we would sign someone up,
(silently) fail to store their passphrase stream, and then they would be out of
luck after returning to the app, unable to login again.

Test have been added that try various secret stores that should be passing
PrimeSecretStores check, but also there are tests which create non-functional
SecretStoreFile instances with read-only backing directory, so we can observe
that this check returns an error when a secret cannot be stored.