folder_block_manager: add a limit on how many ptrs get processed
Controlled by the mode, we can limit how many pointers get archived or
cache-cleaned at once.  If the MD contains too many unref'd pointers,
we don't want to duplicate them all in memory, when memory is
constrained.  This was inspired by an MD in the wild that had 500K+
unref'd pointers.

If doing QR, we can't just reclaim a partial set of pointers from a
single MD, so in that case return an error and let another device
finish the QR for that revision.

Setting the mode in the middle of a test caused data race issues, so I
had to make ConfigLocal.mode sit under a lock.

Issue: KBFS-3984