folder_branch_ops: add debugging for stat'd file size
We saw a problem in the wild where `rm -rf` didn't work.  It's not
clear why, but it seemed like it didn't recurse into a directory after
statting it.  One possible explanation is that KBFS incorrectly
returned a 0 size for the stat'd directory, and that caused `rm` to
skip it even though it had files in it.  This will help us debug it in
the future, if it happens again.

Issue: KBFS-3985