kbfs: vloggify loquacious loggers
This commit turns many of the most prolific log lines into vlog lines,
at debug level 1.  (Prefetcher logs get a demotion to debug level 2.)

As a compromise, at normal debug levels it logs a single line at the
end of each major `folderBranchOps` operation, that includes the
duration of the operation, for performance debugging if needed.  And
if the operation takes more than a second, it logs an opening line
too, to make sure _something_ is logged.

An example end log line is:

```
2019-04-12T17:23:52.666804-07:00 ▶ [DEBU kbfs(FBO 3209b467) folder_branch_ops.go:3014] 16b [duration=220.358µs] GetDirChildren NodeID(0xc421902a20) done, 138 entries: <nil> [tags:FID=7DF4w09kDJkYkrOeYDxMTw]
```

This improves the go source rsync benchmark by about 23%.  If we
removed that single line still being logged, it would improve it by
about 34% instead, but I'm not sure it's worth the loss of
debugability.

This also leaves in place `bserver_remote` and `disk_block_cache` logs
for now, just in case.  But maybe we can remove them in the future.

This also sets the mobile debug level to vlog1, since there is
currently no way to turn on debugging in mobile, and there may be bugs
we can't find without being able to turn this back on.  Since bulk
transfers of data in mobile are unlikely right now, this seems
acceptable.

Issue: KBFS-4054