block_retrieval_queue: don't overwrite non-nil block fetch error
If a block was not being prefetch-tracked, and it encounted an error
while being fetched, we were still trying (and failing) to put the nil
block in the cache.  But the worst part is that we were overriding
`err` to `nil` when we failed to put it back in the cache.  So anyone
waiting on that request just saw an empty block.

One case where we saw this happen is when a block was being fetched on
mobile (where blocks are fetched without prefetching, by default), and
then the syncing was switched on for the TLF during the fetch, causing
the prefetcher to shut down all requests (passing `FinalizeRequest` a
`nil` block).

I think on normal non-prefetcher-shutdown errors (like if the bserver
is unreachable and the context is canceled), the block that gets
passed into `FinalizeRequest` is properly typed, and so putting it in
the cache doesn't fail, and the error isn't overridden.