libgit: add autogit tolerance for submodules
Previously, autogit would EIO if there was a submodule, because it's
represented internally by the commit hash from the other repo.

Instead, detect when we can't find the git object for an entry, and
turn that into a special file that just contains the string "git
submodule at commit ", followed by the commit in the other repo.

It proved nearly impossible to test submodules without running a full
kbfsgit runner process, so I wrote the test under `kbfsgit/` instead
of `libgit/`.  Please have mercy on my soul.

Issue: KBFS-4046