prefetcher: propagate higher priority requests
Partial prefetches work by first doing a normal (non-deep-sync)
lookup, and then a full deep-sync request for the same block ID, at a
high priority.  Any child blocks that get prefetched due to the first
request get dropped to a throttled priority, and when the same blocks
are then prefetched as part of the deep sync, the requests at the
throttle priority are already in the queue and hence still get
throttled.  Which means partial prefetches end up taking forever.

So override the priority in the prefetch queue.

Partial prefetches still seem to take a while, but not an unbounded
amount of time, like before.

And fix a debug message to stringify actions.

Issue: KBFS-4070