prefetcher: check for and handle status requests constantly
`keybase fs show` depends on being able to request the prefetch status
from the prefetcher, which in turn requires sending a message on a
channel and waiting for it to be processed by the main processing
goroutine.  But in some cases that goroutine can be busy for tens of
seconds, e.g. when it is processing a block with 1000s of child
pointers, each one involving a disk cache operation.

So this just processes any pending status requests (which are quick)
inline during every iteration of that child processing, to guarantee a
quick return.  This is safe, since the status request is read-only and
isn't interfering with the prefetching logic at all.