block_retrieval_queue: protect doneCh
There was a race where `Request` checks the doneCh, but and sends the
request to the worker, but the worker is shut down before it can
take the request out of the queue, leading to hangs.

Instead, keep the lock while shutting down the workers, to ensure the
race can't happen.