cr_chains: when identifying types, handle self-contained subdirs
If a directory was both created and deleted within a branch, and a
file within that directory was created, set-attr'd, and deleted within
the branch as well, we had a bug where CR couldn't complete because it
choked on trying to find the deleted file's entry in the deleted
parent block, and the parent block couldn't be fetched from anywhere.

In this case, just assume the file has been deleted, and return early.

Also adds a test that repros the failure.

Issue: KBFS-4114