bserver: add a timeout during authentication
It seems like sometimes in CI, the bserver fails to either receive or
respond to an `Authenticate` request over one TCP connection, or the
message/response is dropped by docker networking.

So give that a timeout that's smaller than the overall test
SyncFromServer timeout, so it doesn't cause tests to fail.

Issue: KBFS-3723
Issue: KBFS-4144