folder_branch_ops: during SyncFromServer, timeout bserver waits
Any action that depends on the bserver to be up could block forever
during the offline tests, so add a timeout to them.  This doesn't
really affect prod behavior, since this function isn't called except
when finalizing git transactions, and in that case these won't have
anything to wait on.

Issue: KBFS-4144