folder_branch_ops: turn a billy.Filesystem EOF into a nil error
When reading from a wrapped node using a `billy.Filesystem`, we get an
EOF error.  But the libkbfs interface promises to return a nil error
in that case (along with 0 bytes read), and if it doesn't, fuse ends
up getting an EIO.  So just catch that case and return a nil error.

Issue: KBFS-4159