tlfhandle: add new handle-making method for const TLF IDs
In testing, where each user has their own `idutil.DaemonLocal`, when a
user creates an iteam TLF ID, they will not propagate it to all users.
Those users will _usually_ be able to look it up via the mdserver, but
in some race situations they will first try to resolve the handle
without doing that, and will end up setting the ID to the null ID.

Instead, if we already have the TLF ID, just make sure it's always
set.  This is a nice cleanup too, avoiding the need for using the
`ConstIDGetter` explicitly when making a handle.

Issue: KBFS-4138