prefetcher: fire new requests if the block refnonce is different
This fixes the following race / test flake:
* A prefetch is started for block A, nonce 0.
* Block A, nonce N is created.
* Block A, nonce 0 is deleted.
* A prefetch is requested for block A, nonce N, but it doesn't trigger
  a fetch because the first request is still in the `prefetches` map.
* The cancel request for block A, nonce 0 is handled by the prefetcher,
  and so the nonce N request is never fulfilled.

This PR fixes the flake by always starting a block request for a new
nonce (i.e., if the BlockPointer is different), to make sure it will
be seen by the prefetcher after the cancel of the original nonce
completes.

Also print out pointers more, instead of IDs, to make debugging things
like this easier.

Issue: KBFS-4161