libfuse: check that .kbfs_status works in subdirs
I had an issue where I couldn't read `.kbfs_status` files in a subdir,
but now I can't reproduce it so might have been a fluke or something
local on my dev machine.  But this seems like a reasonable cleanup of
this test anyway, so might as well commit it.