favorites: adjust TLF mtimes based on the edit history times
To avoid counting GC and git modifications in the favorite mtimes used
to sort folders, cross-check them with the time of of the
corresponding edit history.  If the favorites time is newer, we can
safely set it to the edit history time.  If we don't yet know the edit
history of the TLF, that means it's not in the most recent edit
history list, and we can just set the time to be below the oldest
known time, to keep it below the most-recently-edited TLFs in the
sorted list.

Apparently the favorite mtime is incorrectly in seconds, rather than
in milliseconds, so some math is required.

Issue: KBFS-4200