md_ops: ignore writer-check error if the merkle root is too old
If a TLF was last written by a device that was revoked too long ago
(i.e., the service doesn't support checking its Merkle root because
it's too old), there's nothing much we can do but trust the server.
Note that we already do this for other Merkle checks -- this one was
just missing until now.