Added unit tests
Note that this does not fully test the proxy support since it does not set up a test environment with a
proxy or actually run HTTP requests through a real proxy. This means this is definitely a limited and
imperfect set of tests but it is still definitely better than nothing. As long as go's http.ProxyFromEnvironment
continues to function the same way and we don't switch away from it, things should be reliable.