Fix PICNIC-98 so that when a new user clicks on the wallet tab it takes them to the wallet tab (#18116)
Prior to this commit, when a new user clicked on the wallet tab they would see the introduction/warning
modal but the highlighted tab would stay the same as the one they were previously on. This PR changes it
so that it correctly navigates them to the wallet tab (though the background stays blank). navigateUp
did not appear to work correctly to navigate away from the wallet tab so now when a new user exits out
of the introduction modal they will simply be sent to the first tab (the people tab).