leveldb: on mobile, use smaller write buffer sizes
Each leveldb instance pre-allocates a full write buffer worth of space
when it's opened.  Using 10 MB for each one is wasteful on mobile, so
dial it back to 100 KB, except for the disk block cache.  If the disk
block cache's buffer it too low, it causes too many tables to be
written to disk before compaction, and can result in a "too many open
files" error during a restart.

Issue: HOTPOT-279