libkbfs: load cached MD changes under journal lock
The MDOps interface is usually responsible for putting a completed MD
into the MD cache.  The reason is that the journal needs to do so
under its own lock, to avoid racing with itself when it detects a
conflict, transforms the MD into a branch MD, and replaces it in the
cache.  If something else races with the journal's branch conversion,
it could end up looking like the local MD version is the real MD for
the merged branch, which isn't true when there's a conflict.

There was one place where this was missing, however; in the code that
re-embeds block changes after the MD was put to the server or the
journal.  This code re-embeds the changes, and then replaces the MD in
the cache, which could race with journal conflict detection.

This commit changes the flow so that the block changes are re-embeded
by the MDOps implementation.  This requires passing a `blockPutState`
around to more places, but it guarantees that `folderBranchOps` won't
race against the journal and put an unmerged MD back into the cache as
a merged one.

Issue: HOTPOT-324