Display proxy settings on the login page if the user is not connected to the internet (#18420)
* Display proxy settings on the login page if the user is not connected to the internet

Internet detection is done via the ping endpoint (the same way that `keybase ping` works). If the user is on the join-or-login page and does not have internet, then a link is displayed that when clicked pops open a modal that displays the proxy settings. Since the login page does not support modals (and that work is out of scope for this PR) this modal simply displays on top of the home page. From there users can configure a proxy and save their settings. This allows users to join/login to Keybase even if they only have internet through a proxy.

* Cleaned up PR and ready for review now

* Remove unused imports

* Switch to react native setInterval

* Updated stories

* Switch away from compose in order to ensure we get type checking

* Remove unused imports