Fix airdrop join modal on android
On android the view inside of the scrollview did not take up the full height of the screen.
This led to the buttons overflowing past the end of the view and not being visible. Fixed
by adding `flex: 1` to the scrollview's view via the `contentContainerStyle` prop.

Tested on android and desktop