Display "You can send assets other than XLM" banner in the send form when started via chat messages
This banner did not display since the from field was not getting properly routed to the component. In
order to route this information the wallet icon in chat now calls the loadaccounts RPC in order to load
all of the accounts into the redux store (if they aren't already in the redux store). It then passes the
default account ID to the send form which passes it to the go service when determining whether or not to
show the banner.

Since the RPC is only called when the wallet icon is fist mounted AND the list of accounts in redux is empty,
this is not expected to have significant performance implications. I had initially implemented this as doing
the RPC call once the wallet icon was clicked, but this led to a race condition where the user could quickly
click on the wallet icon and then the send button and open the send form before the load accounts RPC returned.
This race condition is still technically possible but much less likely with this setup.