Add number pad icon [hack warning]
Note that this is a bit hacky. svg2ttf has some issue with this svg where the svgo optimized
image would be rendered by svg2ttf as a black box. By inspecting the uncompressed svg file I
saw that there was a rect over the entire image but without a fill specified. This should not
have been rendered as a black box, but svg2ttf renders it as a blackbox. It turns out svg2ttf
seems to implement its own svg rendering that is subtly broken in this case. Worked around by
just removing the rect and rendering the uncompressed svg file.